Fixed using wrong commit from godot-mono-builds.
Since the mono version used for the containers was updated, the
patches included in the godot-mono-builds repo were outdated, as
the commit checked out was not changed. Changed the commit to the
most recent commit as of time of writing to include up-to-date
patches.

This should not affect the final Tabletop Club builds, this is just
purely so that the build.sh script works the first time around.